chore(i18n): update translations and improve CI i18n checks (#15)
* chore(i18n): update language proficiency levels to C2

Change the proficiency levels for German and English in the
internationalization files to reflect C2 instead of C1.

Update the Transloco configuration to include input paths and
set error handling for extra keys. Modify TypeScript
configuration to include the Transloco config file in the
compilation process.

Reflect changes in GitHub Actions workflow to ensure i18n checks
are run when relevant files are modified.

* chore(tsconfig): update test types to use vitest

Replace jasmine with vitest/globals in the types configuration
for the test TypeScript configuration.